Move endpoint to PATCH /users/{userId}
